# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI 文字游戏引擎 — 游玩、创作、社交全栈 API",
        "title": "WorkshopEngine API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/play/sessions/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除会话及所有关联的楼层、页面和记忆",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "play-engine"
                ],
                "summary": "删除会话",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "修改会话标题、状态或公开标志",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "play-engine"
                ],
                "summary": "更新会话信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新字段",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_engine_api.UpdateSessionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/play/sessions/{id}/branches": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "列出会话所有分支（含 main）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "play-engine"
                ],
                "summary": "列出会话分支",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/play/sessions/{id}/branches/{bid}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除指定分支（不能删除 main 分支）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "play-engine"
                ],
                "summary": "删除分支",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "分支 ID",
                        "name": "bid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/play/sessions/{id}/floors": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "列出会话所有楼层（含激活页摘要），支持 from/to 范围过滤",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "play-engine"
                ],
                "summary": "楼层列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "起始楼层序号",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "结束楼层序号",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/play/sessions/{id}/floors/{fid}/branch": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "从指定楼层创建新分支",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "play-engine"
                ],
                "summary": "从楼层创建分支",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "楼层 ID",
                        "name": "fid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/play/sessions/{id}/floors/{fid}/pages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "列出指定楼层的所有 Swipe 页",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "play-engine"
                ],
                "summary": "Swipe 页列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "楼层 ID",
                        "name": "fid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/play/sessions/{id}/floors/{fid}/pages/{pid}/activate": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "将指定页设为楼层的激活页",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "play-engine"
                ],
                "summary": "激活 Swipe 页",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "楼层 ID",
                        "name": "fid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "页 ID",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/play/sessions/{id}/floors/{fid}/snapshot": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回 Verifier 结果和命中词条",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "play-engine"
                ],
                "summary": "楼层 Prompt 快照",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "楼层 ID",
                        "name": "fid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/play/sessions/{id}/fork": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "从指定楼层分叉出新会话（平行时间线/存档点）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "play-engine"
                ],
                "summary": "分叉会话",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "分叉参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_engine_api.ForkSessionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/play/sessions/{id}/memories": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回会话所有记忆条目",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "play-engine"
                ],
                "summary": "列出记忆条目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "手动创建记忆条目（创作者/调试用）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "play-engine"
                ],
                "summary": "创建记忆",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "记忆内容",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_engine_api.CreateMemoryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/play/sessions/{id}/memories/consolidate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "立即同步触发记忆整合（调试用）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "play-engine"
                ],
                "summary": "触发记忆整合",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_engine_api.CreateMemoryReq": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "importance": {
                    "description": "0–1，默认 0.9",
                    "type": "number"
                },
                "stage_tags": {
                    "description": "阶段标签（空 = 无阶段限制）",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "description": "fact | summary | open_loop，默认 fact",
                    "type": "string"
                }
            }
        },
        "internal_engine_api.ForkSessionReq": {
            "type": "object",
            "properties": {
                "from_floor_seq": {
                    "description": "复制到哪个楼层序号（含）。\nnil = 复制全部已提交楼层；0 = 不复制任何楼层（空历史分叉，相当于同模板新会话）。",
                    "type": "integer"
                },
                "user_id": {
                    "description": "可选：覆盖新 session 的用户 ID",
                    "type": "string"
                }
            }
        },
...
        "internal_engine_api.UpdateSessionReq": {
            "type": "object",
            "properties": {
                "is_public": {
                    "description": "指针类型：nil = 不修改，false/true = 明确设置",
                    "type": "boolean"
                },
                "status": {
                    "description": "active | archived",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}